pyproject.toml
[project]
name = "du1"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.14.1",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
]
